requirements.txt
# 贝叶斯网络虚假评论识别 - 依赖包

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# 文本处理
textblob>=0.17.1
nltk>=3.8.1

# 配置管理
pyyaml>=6.0

# 进度条
tqdm>=4.65.0

# 数据序列化
pyarrow>=12.0.0  # 用于 parquet 格式

